bass_rag/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
rank-bm25>=0.2.2
torch>=2.0.0
transformers>=4.30.0
tiktoken>=0.5.0
influxdb-client>=1.38.0
